requirements.txt
discord.py
vaderSentiment
matplotlib
flask